Update tools.py
added get_fit_isometry which prepares a perspective change matrix that from a set of four points to another set of four points that forms a rectangle with the height and width grabbed from the right left side and top of the quadrilateral formed by the first set of points. not an actual isometry.